promhttp/delta: hold born and renderedGen in 32 bits
Both are Unix seconds, which fit in 32 bits until 2106. born moves next to
kind and fills the four bytes kind was padded out to, which takes the entry
from 120 bytes to 112: out of the 128-byte size class into the 112-byte one,
16 bytes on every instance. Comparisons and genOf widen back to int64.

TestEntryStaysInItsSizeClass fails once a field takes the entry past 112
bytes, so that growing it is a decision rather than a side effect.

memtest (MEMTEST=1): 670 -> 656 bytes per instance, measured together with
the counter change that follows; 641 with both.

+++ prometheus/promhttp/delta/encode_test.go
@@ -175,7 +175,7 @@ func TestFallbackKeepsGeneration(t *testing.T) {
 	w := bufio.NewWriter(&out)
 	enc := expfmt.NewEncoder(w, expfmt.NewFormat(expfmt.TypeTextPlain))
 	if err := encodeFamilies(w, enc, []*dto.MetricFamily{mf}, [][]*entry{rows},
-		func(en *entry) int64 { return en.born }, newEncState(), "gen", ""); err != nil {
+		func(en *entry) int64 { return int64(en.born) }, newEncState(), "gen", ""); err != nil {
 		t.Fatal(err)
 	}
 	w.Flush()
@@ -287,7 +287,7 @@ func TestGenerationLabelIsSorted(t *testing.T) {
 				w := bufio.NewWriter(&out)
 				enc := expfmt.NewEncoder(w, expfmt.NewFormat(expfmt.TypeTextPlain))
 				if err := encodeFamilies(w, enc, []*dto.MetricFamily{mf}, [][]*entry{{{born: 7}}},
-					func(en *entry) int64 { return en.born }, newEncState(), "gen", ""); err != nil {
+					func(en *entry) int64 { return int64(en.born) }, newEncState(), "gen", ""); err != nil {
 					t.Fatal(err)
 				}
 				w.Flush()
@@ -316,7 +316,7 @@ func TestLabelsAreSharedBetweenFamilies(t *testing.T) {
 	w := bufio.NewWriter(&out)
 	enc := expfmt.NewEncoder(w, expfmt.NewFormat(expfmt.TypeTextPlain))
 	es := newEncState()
-	if err := encodeFamilies(w, enc, mfs, rows, func(en *entry) int64 { return en.born }, es, "gen", ""); err != nil {
+	if err := encodeFamilies(w, enc, mfs, rows, func(en *entry) int64 { return int64(en.born) }, es, "gen", ""); err != nil {
 		t.Fatal(err)
 	}
 	w.Flush()
